fix math for distortion

+++ joy_division/src/main.rs
@@ -34,11 +34,11 @@ fn view(app: &App, _model: &Model, frame: Frame) {
         let mut line = Vec::<Vec2>::new();
 
         for j in (STEP..=SIZE - STEP).step_by(STEP) {
-            let distance_to_center = j - SIZE / 2;
-            let variance = (SIZE / 2 - 50 - distance_to_center).max(0) as isize;
-            let random = random::<isize>();
-            let random = random * variance / 2 * -1;
-            let point = pt2(j as f32, i as f32 + random as f32);
+            let distance_to_center = j as f32 - SIZE as f32 / 2.0;
+            let variance = (SIZE as f32 / 2.0 - 50.0 - distance_to_center).max(0.0);
+            let random = random::<f32>();
+            let random = random * variance / 2.0 * -1.0;
+            let point = pt2(j as f32, i as f32 + random);
             line.push(point);
         }
         lines.push(line);
